Add files via upload
Since 1972, the U.S. Census Bureau administered the National Crime Victimization Survey (NCVS) on an annual basis to a selected sample of households from across the US to collect information about personal and household victimization. Its purpose is to collect data on several forms of crime and its consequences which can be later used to develop models for crime comparison and development over time. As such, the NCVS database is a result of this survey administered to a representative sample of the US population and, therefore, constitutes a retrospective observational study.
The dataset used for this paper was derived from the NCVS and contains 8043 observations of 1811 households sampled in 2020. In particular, it contains 42 clusters each consisting of 72 to 337 households proportionally distributed across several demographical variables. Overall, the dataset contains 81 variables which can be divided into three major groups:

(1) Variables used for identification of households (IDHH or ICPSR), time (YEARQ – year and quarter in which the survey was taken), or PANEL_ROT_GROUP which identifies each of the 42 clusters by a unique two-digit code.

(2) Variables used to characterize the demographics of the persons sampled, including gender, race, age, income, education level, employment, and marital status as well as those characterizing the household itself, including urbanicity, gated access, number of vehicles owned, etc.

(3) Variables used to characterize an incident of crime, such as location, whether it was a personal crime (sexual assault, robbery) or a property crime (theft, motor vehicle theft), use of weapons, relationship between the assailant and victim (if the assailant was known), and whether the police was called or not.

The sampling process follows a stratified, multi-stage cluster design. The NCVS sample is drawn from the decennial population census which provides primary sampling units (PSUs). Then a two-stage process is followed to select specific households from these applicable PSUs. First, the US is divided into 1987 units consisting mostly of metropolitan areas or counties so that each first-stage PSU contains at least 7500 persons in the maximum land area of 3000 square miles. These first stage PSUs are then divided to sampling strata. The largest (and therefore most important) PSUs make up their own strata such that they are as homogenous as possible. Secondly, a systematic random sample of housing units is selected from the first-stage PSU strata. By following this sampling strategy – constructing a homogenized strata from which a systematic random sample is drawn – makes for an unbiased sample because each household has an equal chance of being selected.